docs: remove tinyipa references
We have started down the path of eradicating tinyipa usage and testing
because it was bifrucating our contirbutor resources and focus now that
we have also been able to fix some of the CI jobs to be a bit more
scalable.

This does mean we're doing more with dib based images and they are larger,
but we're willing ot pay that tax as a project for consistency and
CI job stability.

Change-Id: I8f96d106a85f6ab4493785e88955196da08af8e9
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>