ci: disable vnc on IPA jobs
The CI jobs used on this pipeline are for IPA. As such, we don't need the VNC/graphical
console services operating on this project pipeline.

Graphical console is facing end users and interacting with a deployed server,
IPA is about deploying the host. There is not a valid flow which would really cross-over
the two because VNC is about connectivity for the instance which is unrelated to IPA.

Change-Id: Ic7fe4983937df8136f20f48f0f2b42d9b7e7bc6d
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>